Initialize: Separate curl initialization from the other parts
The global curl initialization left some memory footprint in case of per thread initialization. First look, it seemed a memory leak, after an investigation it was find out it is the curl global context.

To avoid this misleading footprint, the  S3_initialize() is separated to  two part, a curl init and a light init. The curl init can be called from the main thread once per program start. While the light init can be called from every worker thread.